[UNTESTED] Refactor code for RF24Network library.
Previously, I was developing for network capability using operations
from the RF24 (~data link layer) library, which quickly lead to huge
overhead for establishing a network with multiple nodes and addresses.
Now, I've switched over to using a pre-implemented library that
provides the network layer.

As a result, methods for the socket (which has been renamed from
'interpreter') needed to be completely rewritten, but due to the
network abstraction, their implementations are considerably
cleaner than before.

In addition, the declarations of many structs and buffers have
been relocated.