Network communication working. High failure rate.
Continuing off of the last commit, all network capabilities
have now been fully reimplimented with the RF24Network library,
which is a much cleaner solution. However, sending between
different nodes always fails for some reason. If you send
to the sending node, it always works (though the packets are
received in reverse order). The next step is to figure out
why this is the case.

Note: printf.h was added for more efficient debugging messages.
Also, one no longer uses 's' to broadcast in the interpreter,
but rather, one types the address of the node they wish to
send to. Right now addresses are limited to one digit.