Successful implementation of socket
Included nodeconfig to assign addresses, which will be improved
as a next step. Added an 'E' message, which sends an empty body,
such that the receiver knows when to stop listening for 'L' (line)
messages and to actuate.

Additionaly, there were various buffer overflows and pointer
arithmetic issues that have been corrected.