Add server directory
The server directory contains a python server that accepts HTTP
API requests to control the lights. So far, the API accepts POST
requests to add an entry with a server-assigned index, and accepts
PUT requests to change state. The server connects to the master
node via serial, and the master node sends compiled commands to
worker nodes.

Next is finishing all the code for sending responses. A lingering
problem is that if we want confirmation that the master successfully
sent a command to the worker nodes, we need to read the master
node's response over serial. But, because we do not know the precise
amount of bytes to read, I have just read a large number and added
a timeout (as if there is no timeout, the read() call will block
forever). A 1-sec timeout is slow, but a 0.5-sec timeout is too fast
to clear the welcome message that the master node sends. This will
be addressed in the future