Add entry-entering functionality for frontend
Now, with the use of Javascript commands, the user can
create entries on the screen which represent nodes. There is
a global ENTRY_LIST variable of object type EntryList to which
one adds or deletes entries via ENTRY_LIST.addEntry and
ENTRY_LIST.deleteEntry.

The next step is to allow users to graphically make these
modifications. I will do this by adding event handling to
the UI.