Fix cross-domain issue, change do_PUT to do POST
It turns out that the issue with cross domain validation was
in fact that CHROME, upon handling a call to jQuery.ajax with
method PUT, would translate PUT to OPTIONS, which is not supported
by the web server. Upon further inspection, jQuery documentation
warns that not all browsers support methods other than GET and POST,
although I never suspected Chrome not to support it.

Thus, for the purposes of this experimental code, I made all PUT
requests POST requests, which fixes the issue.