Add synchronization for client--only GUI additions
Added the method fetchServerList(), which repopulates the GUI
with any devices stored on the server via GET /devices/. Note
that currently, because DELETEs are not supported, the GUI does
not delete GUI device elements in order to synchronize them.
That functionality will be added as soon as DELETEs are possible.