First refactor of server side into node framework
The new infrastructure uses an actual database, MongoDB,
along with the Mongoose ODM which is used in the device
model.

Rather than pySerial as in before, we communicate to the
Arduino using the SerialPort node.js module, of which we instantiate
and augment in config/serial.js.

Additionally, the node server serves the GUI HTML controller, meaning
that all internal API calls are within the same domain as the GUI
caller i.e. there are no longer cross-domain calls! Finally! Now
we can use all the HTTP methods we need without worrying about
horrendous preflighting and CORS compliance.