Refactor receive-ACK housekeeping
Now, when the transmitter sends a DIS request, it increments
the fact that it sent a request. When the transmitter receives
an ACK, it logs the round trip time, increments the ACK count,
and increments the total RTT and RTT^2 time (for standard deviation).
This requires the tester to send each test individually, but
provides more reliable analytics and flexbility for different
benchmarks.