Front end functionality for placing nodes on photo
Currently, the angular view allows the user to place nodes
on the map via right-click. The ng-right-click directive
has been implemented to support this. In addition, such points
place on the map are linked to an underlying point object and
linked with a tab at the bottom of the screen.