[WIP] Preliminary work on endpoints for DB interaction
Currently, the front end creates nodes via the graphical user interface
with seperate entries for the front end "point" vs the database's
"node" object. The Point object contains information necessary for
displaying it on the GUI, namely x and y fields, so for every GET
request, the front end lacks values it needs for display.

Next, I will rewrite the API for GETs and POSTs to send JSON rather
than encoding only brightness in the URL. This will keep Points
and Nodes synchronous so the front end can work correctly.