Front-end changes: camera b.ground, node drawing
Note that currently the update node has a bug such that
if there are ever two nodes on the screen with the same
address, the server will delete the least-recently used one,
so for the time being, please ensure that this doesn't happen.
You can do this by setting the nodes' addresses in descending
order.