Refactor front end to lazy upload nodes
Previously, changing the address field on a point's menu caused a
PUT call which deleted the node with the matching address from
the database and readded it. Now, when a front end node is has its
address field changed OR is created, it is only placed in a front
end object that stores nodes.

Then, the user presses the submit button to clear the server and upload
all the new front end nodes. Clearing the server every time may not
scale well, but since we have so few nodes it will be fine for now.

To support this, the API now has PUT /devices which takes a JSON
object of nodes to upload. Additionally, the front end now provides
the nodes with the _id field because we don't immediately make a call
to the server, and thus cannot obtain the Mongo-generated ObjectId _id.