Begin service for distributed behaviors
So far, we have a follow-the-leader behavior. However, for some
reason, the default 'toggle' point behavior does not turn nodes
off after they light up, giving the appearance instead of a
loading bar. But, that's nice as well, so I documented
that as well in the figures.

The next step will be to change follow the leader to work as
expected.