Full refactor into uiRouter
Now, there is a toolbar on the left to navigate between distributed
behaviors, which are represented as different partial views loaded
in via uiRouter. For example, for the click interaction, directives
are added to the _click.html file that enable acuation by clicking,
while with _lineFill, the html elements will (but do not currently as of
this commit) support actuation via the behaviorService.

Note that the there are still some scoping issues, one of which
is with deleteNodes, where $scope.points is still not resolving
to the root scope. A similar problem occured in setting $scope.points,
where multiple definitions in child scopes caused child scopes to
add to empty objects. This was fixed by changing the instantiation
to: $scope.points = $scope.points || {};

This issue may be addressed later by using 'this' instead of scope.