Fully factor paper.js into a UI Router state
Now, once the DOM element for a state is loaded, by means of
$scope.$on('$viewContentLoaded', function() { ... }), we fire
the paper.js script to modify the newly loaded DOM elements.
This business logic is contained within the state's controller.

Note that in the future, we will eventually want to move this
logic outside of the states' definitions.