[WIP] Implement basic node creation, selection, path select
All the logic has been added to the state controller in app.js.
Note that paper.js's style has caused the creation of different
tools, which are in fact bona fide tools in the sense that each
tool contains a different set of mappings from mouse inputs to
effects on the canvas.

Currently, we have a createNode tool and a pathSelect tool. I am
hoping that we can construct enough tools for users to create
behaviors by selecting paths and coordinating individual behaviors.