Complete selection->behavior->block pipeline
Within the sandbox tool, the user now creates nodes with the place
tool, makes a selection with a path tool, presses enter to create
the selection. They then form blocks by pairing an individual
point behavior with a selection. Actuating the block then applies
the behavior to all the nodes in the selection.

Next is organizing the blocks so that one's completion leads
to another