Refactor trigger handling and add onClick()
Now, $scope.triggers is a dictionary mapping trigger names to a
functions of the form f(triggerItem, blockId), where triggerItem
is the selection (so possibly multiple paths/items) serving
as the the triggerObject, and blockId is the block that will
be actuated upon the trigger. F **returns a function** that, gets added
to Paper.js's on____ handlers and contains the triggerItem and blockId
via lexical scoping.

So, upon making a block, we call one of the trigger functions given
the block's trigger and triggerObject, then the trigger function
returns a handler function which will listen for a trigger
ont the triggerObject, and this hander is added to
$scope.triggerHandlers and added to Paper.js's on___ handlers
**for the Interact tool only**. That is, clicking on a trigger
only does anything with the interact tool.

Next is adding the rest of the triggers.